finish refactor of api => context
