+++ api/context.go
@@ -2,10 +2,30 @@ package api
 
 import (
 	"encoding/json"
+	"errors"
+	"fmt"
 	"net/http"
+	"strconv"
+	"strings"
 
 	"github.com/gorilla/mux"
+	. "github.com/mostlygeek/go-debug"
 	"github.com/mostlygeek/go-syncstorage/syncstorage"
+)
+
+var apiDebug = Debug("syncapi")
+
+var (
+	ErrMissingBSOId    = errors.New("Missing BSO Id")
+	ErrInvalidPostJSON = errors.New("Malformed POST JSON")
+)
+
+const (
+	MAX_BSO_PER_POST_REQUEST = 100
+	MAX_BSO_PAYLOAD_SIZE     = 256 * 1024
+
+	// maximum number of BSOs per GET request
+	MAX_BSO_GET_LIMIT = 2500
 )
 
 // NewRouterFromContext creates a mux.Router and registers handlers from
@@ -25,22 +45,19 @@ func NewRouterFromContext(c *Context) *mux.Router {
 
 	info := v.PathPrefix("/info/").Subrouter()
 	info.HandleFunc("/collections", c.hawk(c.uid(c.hInfoCollections))).Methods("GET")
-	/*
-		info.HandleFunc("/quota", c.hawk(c.uid(c.notImplemented))).Methods("GET")
-		info.HandleFunc("/collection_usage", c.hawk(c.uid(c.hInfoCollectionUsage))).Methods("GET")
-		info.HandleFunc("/collection_counts", c.hawk(c.uid(c.hInfoCollectionCounts))).Methods("GET")
+	info.HandleFunc("/collection_usage", c.hawk(c.uid(c.hInfoCollectionUsage))).Methods("GET")
+	info.HandleFunc("/collection_counts", c.hawk(c.uid(c.hInfoCollectionCounts))).Methods("GET")
+	info.HandleFunc("/quota", handleTODO).Methods("GET")
 
-		storage := v.PathPrefix("/storage/").Subrouter()
-		storage.HandleFunc("/", c.hawk(c.uid(c.notImplemented))).Methods("DELETE")
+	storage := v.PathPrefix("/storage/").Subrouter()
+	storage.HandleFunc("/", handleTODO).Methods("DELETE")
 
-		storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionGET))).Methods("GET")
-		storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionPOST))).Methods("POST")
-		storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionDELETE))).Methods("DELETE")
-
-		storage.HandleFunc("/{collection}/{bsoId}", c.hawk(c.uid(c.hBsoGET))).Methods("GET")
-		storage.HandleFunc("/{collection}/{bsoId}", c.hawk(c.uid(c.hBsoPUT))).Methods("PUT")
-		storage.HandleFunc("/{collection}/{bsoId}", c.hawk(c.uid(c.hBsoDELETE))).Methods("DELETE")
-	*/
+	storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionGET))).Methods("GET")
+	storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionPOST))).Methods("POST")
+	storage.HandleFunc("/{collection}", c.hawk(c.uid(c.hCollectionDELETE))).Methods("DELETE")
+	storage.HandleFunc("/{collection}/{bsoId}", c.hawk(c.uid(c.hBsoGET))).Methods("GET")
+	storage.HandleFunc("/{collection}/{bsoId}", c.hawk(c.uid(c.hBsoPUT))).Methods("PUT")
+	storage.HandleFunc("/{collection}/{bsoId}", handleTODO).Methods("DELETE")
 
 	return r
 }
@@ -104,6 +121,39 @@ func (c *Context) JSON(w http.ResponseWriter, r *http.Request, val interface{}) 
 	}
 }
 
+// getcid turns the collection name in the URI to its internal Id number. the `automake`
+// parameter will auto-make it if it doesn't exist
+func (c *Context) getcid(r *http.Request, uid string, automake bool) (cId int, err error) {
+	collection := mux.Vars(r)["collection"]
+
+	if !syncstorage.CollectionNameOk(collection) {
+		err = syncstorage.ErrInvalidCollectionName
+		return
+	}
+
+	cId, err = c.Dispatch.GetCollectionId(uid, collection)
+
+	if err == nil {
+		return
+	}
+
+	if err == syncstorage.ErrNotFound {
+		if automake {
+			cId, err = c.Dispatch.CreateCollection(uid, collection)
+		}
+	}
+
+	return
+}
+
+// Ok writes a 200 response with a simple string body
+func okResponse(w http.ResponseWriter, s string) {
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("X-Content-Type-Options", "nosniff")
+	w.WriteHeader(http.StatusOK)
+	fmt.Fprint(w, s)
+}
+
 func (c *Context) handleHeartbeat(w http.ResponseWriter, r *http.Request) {
 	// todo check dependencies to make sure they're ok..
 	okResponse(w, "OK")
@@ -121,3 +171,325 @@ func (c *Context) hInfoCollections(w http.ResponseWriter, r *http.Request, uid s
 		c.JSON(w, r, info)
 	}
 }
+
+func (c *Context) hInfoCollectionUsage(w http.ResponseWriter, r *http.Request, uid string) {
+	results, err := c.Dispatch.InfoCollectionUsage(uid)
+	if err != nil {
+		c.Error(w, r, err)
+	} else {
+		c.JSON(w, r, results)
+	}
+}
+
+func (c *Context) hInfoCollectionCounts(w http.ResponseWriter, r *http.Request, uid string) {
+	results, err := c.Dispatch.InfoCollectionCounts(uid)
+	if err != nil {
+		c.Error(w, r, err)
+	} else {
+		c.JSON(w, r, results)
+	}
+}
+
+func (c *Context) hCollectionGET(w http.ResponseWriter, r *http.Request, uid string) {
+
+	// query params that control searching
+	var (
+		err    error
+		ids    []string
+		newer  int
+		full   bool
+		limit  int
+		offset int
+		sort   = syncstorage.SORT_NEWEST
+	)
+
+	cId, err := c.getcid(r, uid, false)
+
+	if err != nil {
+		if err == syncstorage.ErrNotFound {
+			w.Header().Set("Content-Type", "application/json")
+			w.Write([]byte("[]"))
+			return
+		} else {
+			c.Error(w, r, err)
+			return
+
+		}
+	}
+
+	if err = r.ParseForm(); err != nil {
+		http.Error(w, "Bad query parameters", http.StatusBadRequest)
+		return
+	}
+
+	if v := r.Form.Get("ids"); v != "" {
+		ids = strings.Split(v, ",")
+		for i, id := range ids {
+			id = strings.TrimSpace(id)
+			if syncstorage.BSOIdOk(id) {
+				ids[i] = id
+			} else {
+				http.Error(w, fmt.Sprintf("Invalid bso id %s", id), http.StatusBadRequest)
+				return
+			}
+		}
+
+		if len(ids) > 100 {
+			http.Error(w, fmt.Sprintf("Too many ids provided"), http.StatusRequestEntityTooLarge)
+			return
+		}
+	}
+
+	// we expect to get sync's two decimal timestamps, these need
+	// to be converted to milliseconds
+	if v := r.Form.Get("newer"); v != "" {
+		floatNew, err := strconv.ParseFloat(v, 64)
+		if err != nil {
+			http.Error(w, "Invalid newer param format", http.StatusBadRequest)
+			return
+		}
+
+		newer = int(floatNew * 1000)
+		if !syncstorage.NewerOk(newer) {
+			http.Error(w, "Invalid newer value", http.StatusBadRequest)
+			return
+		}
+	}
+
+	if v := r.Form.Get("full"); v != "" {
+		full = true
+	}
+
+	if v := r.Form.Get("limit"); v != "" {
+		limit, err = strconv.Atoi(v)
+		if err != nil || !syncstorage.LimitOk(limit) {
+			http.Error(w, "Invalid limit value", http.StatusBadRequest)
+			return
+		}
+	}
+
+	// assign a default value for limit if nothing is supplied
+	if limit == 0 {
+		if c.MaxBSOGetLimit > 0 { // only use this if it was set
+			limit = c.MaxBSOGetLimit
+		} else {
+			limit = MAX_BSO_GET_LIMIT
+		}
+	}
+
+	// make sure limit is smaller than c.MaxBSOGetLimit if it is set
+	if limit > c.MaxBSOGetLimit && c.MaxBSOGetLimit > 0 {
+		limit = c.MaxBSOGetLimit
+	}
+
+	// finally a global max that we never want to go over
+	// TODO is this value ok for prod?
+	if limit > MAX_BSO_GET_LIMIT {
+		limit = MAX_BSO_GET_LIMIT
+	}
+
+	if v := r.Form.Get("offset"); v != "" {
+		offset, err = strconv.Atoi(v)
+		if err != nil || !syncstorage.OffsetOk(offset) {
+			http.Error(w, "Invalid offset value", http.StatusBadRequest)
+			return
+		}
+	}
+
+	if v := r.Form.Get("sort"); v != "" {
+		switch v {
+		case "newest":
+			sort = syncstorage.SORT_NEWEST
+		case "oldest":
+			sort = syncstorage.SORT_OLDEST
+		case "index":
+			sort = syncstorage.SORT_INDEX
+		default:
+			http.Error(w, "Invalid sort value", http.StatusBadRequest)
+			return
+		}
+	}
+
+	results, err := c.Dispatch.GetBSOs(uid, cId, ids, newer, sort, limit, offset)
+	if err != nil {
+		c.Error(w, r, err)
+		return
+	}
+
+	w.Header().Set("X-Weave-Records", strconv.Itoa(results.Total))
+	if results.More {
+		w.Header().Set("X-Weave-Next-Offset", strconv.Itoa(results.Offset))
+	}
+
+	if full {
+		c.JSON(w, r, results.BSOs)
+	} else {
+		bsoIds := make([]string, len(results.BSOs))
+		for i, b := range results.BSOs {
+			bsoIds[i] = b.Id
+		}
+		c.JSON(w, r, bsoIds)
+	}
+}
+
+func (c *Context) hCollectionPOST(w http.ResponseWriter, r *http.Request, uid string) {
+	// accept text/plain from old (broken) clients
+	if ct := r.Header.Get("Content-Type"); ct != "application/json" && ct != "text/plain" {
+		http.Error(w, "Not acceptable Content-Type", http.StatusUnsupportedMediaType)
+		return
+	}
+
+	// parsing the results is sort of ugly since fields can be left out
+	// if they are not to be submitted
+	var posted syncstorage.PostBSOInput
+
+	decoder := json.NewDecoder(r.Body)
+	err := decoder.Decode(&posted)
+	if err != nil {
+		http.Error(w, "Invalid JSON posted", http.StatusBadRequest)
+		return
+	}
+
+	if len(posted) > MAX_BSO_PER_POST_REQUEST {
+		http.Error(w, fmt.Sprintf("Exceeded %d BSO per request", MAX_BSO_PER_POST_REQUEST),
+			http.StatusRequestEntityTooLarge)
+		return
+	}
+
+	// validate basic bso data
+	for _, b := range posted {
+		if !syncstorage.BSOIdOk(b.Id) {
+			http.Error(w, "Invalid or missing Id in data", http.StatusBadRequest)
+			return
+		}
+
+		if b.Payload != nil && len(*b.Payload) > MAX_BSO_PAYLOAD_SIZE {
+			http.Error(w, fmt.Sprintf("%s payload greater than max of %d bytes",
+				b.Id, MAX_BSO_PAYLOAD_SIZE), http.StatusBadRequest)
+			return
+		}
+	}
+
+	cId, err := c.getcid(r, uid, true) // automake the collection if it doesn't exit
+	if err != nil {
+		if err == syncstorage.ErrInvalidCollectionName {
+			http.Error(w, err.Error(), http.StatusBadRequest)
+		} else {
+			c.Error(w, r, err)
+		}
+		return
+	}
+
+	results, err := c.Dispatch.PostBSOs(uid, cId, posted)
+	if err != nil {
+		c.Error(w, r, err)
+	} else {
+		c.JSON(w, r, results)
+	}
+}
+
+func (c *Context) hCollectionDELETE(w http.ResponseWriter, r *http.Request, uid string) {
+
+	cId, err := c.getcid(r, uid, false)
+	if err == nil {
+		err = c.Dispatch.DeleteCollection(uid, cId)
+	}
+
+	if err != nil {
+		if err != syncstorage.ErrNotFound {
+			c.Error(w, r, err)
+		}
+	} else {
+		w.Header().Set("Content-Type", "text/plain")
+		w.Write([]byte("ok"))
+	}
+}
+
+func (c *Context) hBsoGET(w http.ResponseWriter, r *http.Request, uid string) {
+
+	bId, ok := mux.Vars(r)["bsoId"]
+	if !ok || !syncstorage.BSOIdOk(bId) {
+		http.Error(w, "Invalid bso ID", http.StatusBadRequest)
+		return
+	}
+
+	var (
+		cId int
+		err error
+		bso *syncstorage.BSO
+	)
+
+	cId, err = c.getcid(r, uid, false)
+	if err == nil {
+		bso, err = c.Dispatch.GetBSO(uid, cId, bId)
+		if err == nil {
+			c.JSON(w, r, bso)
+			return
+		}
+	}
+
+	if err == syncstorage.ErrNotFound {
+		http.NotFound(w, r)
+		return
+	} else {
+		c.Error(w, r, err)
+	}
+}
+
+func (c *Context) hBsoPUT(w http.ResponseWriter, r *http.Request, uid string) {
+
+	bId, ok := mux.Vars(r)["bsoId"]
+	if !ok || !syncstorage.BSOIdOk(bId) {
+		http.Error(w, "Invalid bso ID", http.StatusBadRequest)
+		return
+	}
+
+	var (
+		cId      int
+		modified int
+		err      error
+	)
+
+	var bso struct {
+		Payload   *string `json:"payload"`
+		SortIndex *int    `json:"sortindex"`
+		TTL       *int    `json:"ttl"`
+	}
+
+	cId, err = c.getcid(r, uid, false)
+	if err == syncstorage.ErrNotFound {
+		http.NotFound(w, r)
+		return
+	}
+
+	decoder := json.NewDecoder(r.Body)
+	err = decoder.Decode(&bso)
+
+	if err != nil {
+		http.Error(w, "Invalid JSON", http.StatusBadRequest)
+		return
+	}
+
+	modified, err = c.Dispatch.PutBSO(uid, cId, bId, bso.Payload, bso.SortIndex, bso.TTL)
+
+	if err != nil {
+		if err == syncstorage.ErrPayloadTooBig {
+			http.Error(w, http.StatusText(http.StatusRequestEntityTooLarge), http.StatusRequestEntityTooLarge)
+			return
+		}
+
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	m := syncstorage.ModifiedToString(modified)
+	w.Header().Set("Content-Type", "text/plain")
+	w.Header().Set("X-Last-Modified", m)
+	w.Write([]byte(m))
+}
+
+/*
+func hBsoDELETE(w http.ResponseWriter, r *http.Request, d *Dependencies, uid string) {
+	notImplemented(w, r, d, uid)
+}
+*/
